refactor-ui-43-섹션-타이틀-설명-표준화
- Section.jsx: title/titleDescription props 추가
- 기존 right prop 제거 및 전역 치환(titleDescription)
- HomePage에 섹션 제목/설명 적용, 간격 통일
