+++ .github/ISSUE_TEMPLATE/ISSUE_TEMPLATE.md
@@ -1,0 +1,18 @@
+## 배경
+
+>
+
+## 작업 항목
+
+- [ ]
+- [ ]
+- [ ]
+
+## 브랜치명
+
+`TYPE-SCOPE-NUMBER-VERB--`
+
+## ✅ 시작 전, 체크!
+
+- [ ] ISSUE-NUMBER-? = #
+- [ ] same-NUMBER-with-BRANCH-?
