+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,14 @@
+## 변경 내용
+
+-
+-
+-
+
+## 체크리스트
+
+- [ ] 로컬 빌드/테스트 통과
+- [ ] UI 확인(필요시 스샷)
+
+## 스크린샷
+
+## Closes
