+++ src/components/common/index.js
@@ -1,7 +1,7 @@
+export { default as NavigationBar } from "../layout/NavigationBar.jsx";
 export { default as Button } from "./Button.jsx";
 export { default as Container } from "./Container.jsx";
 export { default as Image } from "./Image.jsx";
-export { default as NavigationBar } from "./NavigationBar.jsx";
 export { default as Section } from "./Section.jsx";
 export { default as Skeleton } from "./Skeleton.jsx";
 // export { default as ThemeToggle } from "./ThemeToggle.jsx"; // 버튼 숨김이면 그대로 export
